+++ src/FenceTurtleMotion.h
@@ -10,7 +10,7 @@ public:
 	FenceTurtleMotion();
 	~FenceTurtleMotion();
 
-	TurtleMove::const_iterator * computeDestination(Point2D actual_position, float direction, float distance, WindowSpec window) const;
+	TurtleMove::const_iterator * computeDestination(Point2D actual_position, double direction, double distance, WindowSpec window) const;
 	int type() const { return FENCE_MODE; }
 };
