+++ src/TurtleMotion.h
@@ -13,7 +13,7 @@ class TurtleMotion
 {
 public:
 	virtual ~TurtleMotion() {};
-	virtual TurtleMove::const_iterator * computeDestination(Point2D actual_position, float direction, float distance, WindowSpec window) const = 0;
+	virtual TurtleMove::const_iterator * computeDestination(Point2D actual_position, double direction, double distance, WindowSpec window) const = 0;
 	virtual int type() const = 0;
 };
