+++ src/WrapTurtleMotion.h
@@ -10,7 +10,7 @@ public:
 	WrapTurtleMotion();
 	~WrapTurtleMotion();
 
-	TurtleMove::const_iterator * computeDestination(Point2D actual_position, float direction, float distance, WindowSpec window) const;
+	TurtleMove::const_iterator * computeDestination(Point2D actual_position, double direction, double distance, WindowSpec window) const;
 	int type() const { return WRAP_MODE; }
 
 private:
